% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/absmax.r
\name{absmax}
\alias{absmax}
\title{Maximum absolute value of a vector}
\usage{
absmax(x = NULL, return_abs = TRUE)
}
\arguments{
\item{x}{vector of numbers. This is a required element.}

\item{return_abs}{Logical scalar; if TRUE, return the absolute maximum value.
If FALSE, maintain the sign of the original value.
Default is TRUE.}
}
\value{
Maximum of the absolute value of a vector
}
\description{
Generates the maximum of the absolute value of a numeric vector.
}
\details{
Version 2.1, 2026-05-26
}
\examples{
r50 <- rnorm(50)
r50.absmax <- absmax(r50)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
