% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_outlier_lines.r
\name{add_outlier_lines}
\alias{add_outlier_lines}
\title{Plot table from X-13ARIMA-SEATS seasonal adjustment run.}
\usage{
add_outlier_lines(
  seas_obj = NULL,
  otl_col = NULL,
  add_sub_title = TRUE,
  sub_title_cex = 0.7
)
}
\arguments{
\item{seas_obj}{A seas object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{otl_col}{Character array of length 6; color used for different outliers,
with the order being \code{'ao','ls','tc','so','rp','tls'}.
Default is colors from the \code{RColorBrewer} palette \code{"Paired"}
\insertCite{RColorBrewer}{blstinyplot}.}

\item{add_sub_title}{Logical scalar; indicates if plots will include
subtitles denoting what series are plotted.
Default includes a subheader.}

\item{sub_title_cex}{Numeric scalar; scaling for main title of plot.
Default is \code{0.7}.}
}
\value{
Outlier lines are added to an existing \code{tinyplot} plot.
}
\description{
Generate plot of user-specified table from the output of a \code{seas} object generated by the \code{seasonal} package.
}
\details{
Version 1.3, 2026-04-08
}
\examples{
shoes_seas <-
    seasonal::seas(shoes2008, x11 = "", 
       transform.function = "log",
       regression.aictest = c("td", "easter"),
       outlier.types = "all",
       arima.model = "(0 1 1)(0 1 1)",
       forecast.maxlead = 60)
shoes_sa <- seasonal::final(shoes_seas)
plot_series(shoes_sa, 
     y_axis_label = "Shoes", 
     do_grid = TRUE,
     draw_recess = TRUE, 
     this_col = "black",
     start_plot = c(1958,1), 
     this_point_type = 1,
     main_title = "X-11 Adjustment for Shoe Store Sales",
     main_title_line = 1.5,
     main_title_cex = 1.0, 
     this_reset = TRUE)
add_outlier_lines(shoes_seas)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
