% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_color_codes.R
\name{cnv_color_codes}
\alias{cnv_color_codes}
\title{Generates color names from hexidecimal input}
\usage{
cnv_color_codes(color_vec = NULL)
}
\arguments{
\item{color_vec}{vector of color codes
This is a required entry.}
}
\value{
vector of color names closest to hexidecimal color input.
}
\description{
Generates vector of closest color names from hexidecimal color input.
This uses the \code{color.id} function of the \code{plotrix} package
\insertCite{plotrix}{blstinyplot} to do the conversion.
}
\details{
Version 1.1, 2026-04-01
}
\examples{
Moonrise_Codes <- 
   c("#F3DF6C", "#CEAB07", "#D5D5D3", "#24281A", "#798E87", 
     "#C27D38", "#CCC591", "#29211F", "#85D4E3", "#F4B5BD", 
     "#9C964A", "#CDC08C", "#FAD77B")
Moonrise_All <- cnv_color_codes(Moonrise_Codes)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
