% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color_blind_palette.r
\name{color_blind_palette}
\alias{color_blind_palette}
\title{Color-blind friendly color palette}
\usage{
color_blind_palette(with_grey = TRUE)
}
\arguments{
\item{with_grey}{Logical scalar; whether color blind pallate contains \code{'grey'}, otherwise the palette contains \code{black}. Default is TRUE.}
}
\value{
vector of hexadecimal color codes that form a color palette that can be distinguished by color-blind people.
}
\description{
Color palettes that can be used that can be distinguished by color-blind people
(either from \insertCite{RColorBrewer;textual}{blstinyplot}
or \insertCite{cookbookrColorsggplot2;textual}{blstinyplot}).
}
\details{
Version 1.1, 2026-04-01
}
\examples{
this_color_blind <- color_blind_palette(FALSE)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
