% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_to_tis.r
\name{convert_date_to_tis}
\alias{convert_date_to_tis}
\title{Convert ts dates to tis format}
\usage{
convert_date_to_tis(
  this_date,
  this_freq = 12,
  is_start = TRUE,
  return_tis = FALSE
)
}
\arguments{
\item{this_date}{numeric scalar or vector; \code{ts} date to be converted}

\item{this_freq}{numeric scalar; frequency of \code{ts} time series.
Default is 12 (monthly).}

\item{is_start}{logical scalar; is date assumed to be the beginning of the month?
Default is TRUE; if FALSE, date is assumed to be at the
end of the month.}

\item{return_tis}{logical scalar; If true, return as \code{tis} object;
otherwise return as integer.
Default is FALSE.}
}
\value{
A \code{tis} index value that is the equivalent of the
\code{ts} date.
}
\description{
Convert dates used for monthly (or quarterly) \code{ts} series
to \code{tis} \insertCite{tis2021}{blstinyplot} formats.
}
\details{
Version 1.0, 2026-04-01
}
\examples{
this_month <- as.numeric(substr(Sys.Date(),1,4)) +
    (as.numeric(substr(Sys.Date(),6,7)) - 1) / 12
end_this_month_tis <- 
    convert_date_to_tis(this_month, this_freq = 12,
                        is_start = FALSE, return_tis = TRUE)
start_this_month_tis <- 
    convert_date_to_tis(this_month, this_freq = 12,
                        is_start = TRUE, return_tis = TRUE)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
