% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_color.r
\name{display_color}
\alias{display_color}
\title{Color name for display}
\usage{
display_color(this_color_code = NULL, strip_numbers = FALSE)
}
\arguments{
\item{this_color_code}{character string of color code to be used in plot
This is a required entry.}

\item{strip_numbers}{logical scalar that controls if numbers at
the end of the text are stripped from the color name.
Default is TRUE.}
}
\value{
character string of color name closest to hexidecimal color input
(if used) stripped of numbers if \code{strip_numbers = TRUE}.
}
\description{
Generates color names for display on plot labels and subheaders.
}
\details{
Version 1.1, 2026-04-07
}
\examples{
this_color_blind <- color_blind_palette()
this_color_blind_text <- array(NA, dim = 8)
for (i in 1:8) {
    this_color_blind_text[i] <- display_color(this_color_blind[i])
}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
