% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_recession.R
\name{draw_recession}
\alias{draw_recession}
\title{Draw NBER recessions}
\usage{
draw_recession(
  this_col_recess = NULL,
  this_density = 50,
  this_border = NA,
  this_add_recess_start = NULL,
  this_sub_recess = TRUE,
  this_sub_line = 2.5,
  this_sub_cex = 0.75,
  display_color_strip = FALSE
)
}
\arguments{
\item{this_col_recess}{Character string; color used for shading recession periods.
Default is \code{'lightgrey'}.}

\item{this_density}{the density of shading lines, in lines per inch.
The default value is \code{50}. A zero value of density means
no shading lines whereas negative values (and NA) suppress shading
(and so allow color filling).}

\item{this_border}{Integer scalar; thickness of border around region.
Default is \code{NA}, meaning the border is not generated.}

\item{this_add_recess_start}{numeric scalar; Starting date for
an additional recession period at the end of the series.
Default is not to add recession dates.}

\item{this_sub_recess}{Logical scalar; indicates if x-axis label
for recession is produced for this plot.
Default is x-axis label is produced}

\item{this_sub_line}{Integer scalar; position of subtitle of plot.
Default is \code{2.5}.}

\item{this_sub_cex}{Numeric scalar; scaling for subtitle of plot.
Default is \code{0.75}.}

\item{display_color_strip}{Logical scalar; indicates if the display color will be stripped of trailing numbers. Default is FALSE.}
}
\value{
Shades recession dates in plots generated by \code{tinyplot}
\insertCite{tinyplot}{blstinyplot}.
}
\description{
Draws shaded areas in plots corresponding to NBER recessions
\insertCite{nber}{blstinyplot}.
}
\details{
Version 1.0, 2026-04-01
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, 
       arima.model = '(0 1 1)(0 1 1)', 
       x11='')
plot_table(air_seas, 'd12', 'AirPassengers', 
    do_grid = TRUE, 
    draw_recess = FALSE,
    use_ratio = FALSE, 
    add_sub_title = TRUE, 
    this_col = 'forestgreen')
draw_recession(this_col_recess = 'lightblue', 
    this_border = 1, 
    this_sub_line = 2.25, 
    this_sub_cex = 0.75)
}
