% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_peak.R
\name{flag_peak}
\alias{flag_peak}
\title{Flag visual significant peaks in spectra}
\usage{
flag_peak(seas_obj = NULL, spec_type = NULL, spec_freq_code = NULL)
}
\arguments{
\item{seas_obj}{\code{seas} object generated from a call of the \code{seas}
of the \code{seasonal} package on a single time series.
This is a required argument.}

\item{spec_type}{Character string; type of spectrum.
Possible values are \code{'ori'}, \code{'irr'}, \code{'rsd'}, \code{'sa'}.}

\item{spec_freq_code}{Character string; type of frequency being tested.
Possible values are \code{'s'} or \code{'t'}.}
}
\value{
If visually significant peaks found, a numeric vector of the
position of the peak frequecies. If no peaks found, 0.
}
\description{
Determine positions of visual significant peaks in spectra using
the criteria of \insertCite{Soukup99;textual}{blstinyplot}.
}
\details{
Version 1.0, 2026-04-07
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, arima.model = '(0 1 1)(0 1 1)', x11='')
this_flagged_peak_seas <- flag_peak(air_seas,'ori','s')
this_flagged_peak_td <- flag_peak(air_seas,'ori','t')
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
