% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_tick_positions.r
\name{gen_tick_positions}
\alias{gen_tick_positions}
\title{Generate tick mark positions}
\usage{
gen_tick_positions(
  this_series = NULL,
  tick_interval = 2,
  is_date = FALSE,
  time_interval = "years",
  this_base = 1
)
}
\arguments{
\item{this_series}{numeric or date vector, the values used for the x-axis.
This is a required entry.}

\item{tick_interval}{integer scalar; the value separating the tick marks.
Default is \code{2}.}

\item{is_date}{logical scalar; indicates if \code{this_series} is a vector
of \code{Date} objects.
Default is \code{FALSE}, \code{this_series} is numeric.}

\item{time_interval}{character scalar; the time unites used to construct
the x-axis if \code{this_series} is a \code{Date} object vector.
Default is \code{"years"}.}

\item{this_base}{numeric scalar; number used to integer divide \code{this_series}
for numeric vectors.
Default is \code{1}.}
}
\value{
Generates values for major tick marks for the X and Y axis of plots.
}
\description{
Generate tick mark position for x or y axis for base R plots
}
\details{
Version 1.2, 2026-04-06
}
\examples{
air_time <- time(AirPassengers)
air_Date <- 
   as.Date(paste(as.numeric(floor(time(AirPassengers))), 
           as.numeric(cycle(AirPassengers)), 
           "1", sep = "-"))

air_x_axis_time <- 
   gen_tick_positions(air_time, tick_interval = 3)
air_x_axis_Date <- 
   gen_tick_positions(air_Date, tick_interval = 3, is_date = TRUE)
air_y_axis      <- 
   gen_tick_positions(AirPassengers, tick_interval = 200, 
                      this_base = 100)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
