% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recession_dates.R
\name{get_recession_dates}
\alias{get_recession_dates}
\title{Get NBER recession dates}
\usage{
get_recession_dates(
  start_recess = NULL,
  end_recess = NULL,
  add_recess_start = NULL,
  this_freq = 12
)
}
\arguments{
\item{start_recess}{numeric scalar; Starting date for plot.
Default is first recession starting date.}

\item{end_recess}{numeric scalar; Ending date for plot.
Default is last recession ending date.}

\item{add_recess_start}{numeric scalar; Starting date for an
additional recession period at the end of the series.
Default is not to add recession dates.}

\item{this_freq}{numeric scalar; frequency of \code{ts} time series.
Default is \code{12} (monthly).}
}
\value{
Starting and ending dates for NBER recessions within a span of data
}
\description{
Add shading for US NBER recession dates \insertCite{nber}{blstinyplot}.
Recession dates are taken from the \code{nberDates} function of the
\code{tis} package \insertCite{tis2021}{blstinyplot}.
}
\details{
Version 1.0, 2026-04-01
}
\examples{
plot_limits <- c(1949, 1961)
thisRec <-
   get_recession_dates(start_recess = plot_limits[1], 
                       end_recess = plot_limits[2])
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
