% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reg_string.r
\name{get_reg_string}
\alias{get_reg_string}
\title{Get names of regressors}
\usage{
get_reg_string(seas_obj = NULL, xreg_names = NULL, exclude_user = FALSE)
}
\arguments{
\item{seas_obj}{A seas object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{xreg_names}{Character vector with names of user defined regressors used in model.
Default is \code{NULL}, no user defined regressors.
Number of names in this vector should match number of user-defined regressors;
if not, a warning message will be produced.}

\item{exclude_user}{Logical scalar; determines if user defined regressors are
included in the string of regression names output by the function.
Default is \code{FALSE}}
}
\value{
Character string containing a summary of the regressors in the regARIMA model.
If no regressors in the model, the routine will return a blank character.
}
\description{
Generate string of names for the regressors used in the model fit for a given series.
}
\details{
Version 1.1, 2026-04-07
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, 
                  transform.function = "log", 
                  forecast.maxlead=36, 
                  check.print = c( "pacf", "pacfplot" ), 
                  x11 = "", slidingspans = "")
air_reg <- get_reg_string(air_seas)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
