% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_udg_index.r
\name{get_udg_index}
\alias{get_udg_index}
\title{Index for entry in UDG list}
\usage{
get_udg_index(udg_list = NULL, this_key = NULL)
}
\arguments{
\item{udg_list}{List object generated by \code{udg} function of the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{this_key}{Keyword found in \code{udg} files generated by X-13ARIMA-SEATS.
This is a required entry.}
}
\value{
An integer denoting which element in the \code{udg} output matches
the key provided by the user. If there is no match, the function
returns the number \code{0}.
}
\description{
Return index for entry in UDG list returned by the \code{udg()} function of the
\code{seasonal} package.
}
\details{
Version 3.1, 2026-03-05
}
\examples{
ukgas_seas <- 
   seasonal::seas(UKgas, 
      series.period = 4, 
      transform.function = "log", 
      arima.model = "(0 1 1)(0 1 1)", 
      forecast.maxlead = 20,
      check.print = c( "pacf", "pacfplot" ), 
      x11 = "")
ukgas_udg <- seasonal::udg(ukgas_seas)
ukgas_udg_index_a <- 
   get_udg_index(ukgas_udg, this_key='a')
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
