% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcst.r
\name{plot_fcst}
\alias{plot_fcst}
\title{Forecast plot}
\usage{
plot_fcst(
  seas_obj = NULL,
  main_title = "ARIMA forecasts",
  do_sub = TRUE,
  do_grid = FALSE,
  do_frame = TRUE,
  length_ori = 2,
  this_col = c("darkgrey", "blue", "darkgreen"),
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_main_cex = 1,
  this_sub_cex = 0.7,
  this_axis_cex = NULL,
  x_axis_label = "Time",
  y_axis_label = NULL,
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  main_title_line = 2.25,
  sub_title_line = 0.75,
  this_reset = TRUE
)
}
\arguments{
\item{seas_obj}{A seas object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{main_title}{Character string; main title of plot.
Default is  \code{'ARIMA Residuals'}.}

\item{do_sub}{Logical scalar; indicates if subtitle is generated.
Default is to generate the subtitle.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_frame}{Logical scalar; indicates the plot will have a frame.
Default is the plot frame will be produced.}

\item{length_ori}{Integer scalar; number of years of the original series to show
with forecasts. Default is 2 years.}

\item{this_col}{Array of character strings; color used for original series,
forecast, and forecast bounds.
Default is \code{c("darkgrey", "blue", "darkgreen")}.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_main_cex}{Numeric scalar; scaling for main plot title.
Default is the value of \code{1.0}.}

\item{this_sub_cex}{Numeric scalar; scaling for plot subtitle.
Default is the value of \code{0.7}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{x_axis_label}{Character string; x-axis label for plot, if specified.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2.25}.}

\item{sub_title_line}{Integer scalar; position of main title of plot.
Default is \code{0.75}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generates a plot of the regARIMA forecasts with confidence bounds.
}
\description{
Generates regARIMA forecasts with confidence bounds
}
\details{
Version 1.5, 2026-05-15
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, 
                  arima.model = '(0 1 1)(0 1 1)', 
                  forecast.maxlead = 60, 
                  forecast.save = "fct", 
                  series.save = "a1")
plot_fcst(air_seas, 
          main_title = 'Forecasts for Airline Passengers', 
          do_grid = TRUE)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
