% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fcst_history.r
\name{plot_fcst_history}
\alias{plot_fcst_history}
\title{Generate forecast history plot}
\usage{
plot_fcst_history(
  seas_mdl1 = NULL,
  seas_mdl2 = NULL,
  start_hist = NULL,
  main_title = "Differences in the Sum of Squared Forecast Errors",
  name_mdl1 = "Model 1",
  name_mdl2 = "Model 2",
  do_grid = FALSE,
  do_frame = TRUE,
  this_col = NULL,
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_main_cex = NULL,
  this_sub_cex = NULL,
  this_axis_cex = NULL,
  this_mar = c(5.1, 2.1, 5.1, 0.5),
  main_title_line = 2.5,
  sub_title_line = 1,
  display_color_strip = TRUE,
  x_axis_label = NULL,
  y_axis_label = NULL,
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  this_reset = TRUE
)
}
\arguments{
\item{seas_mdl1}{A \code{seas} object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot} for the
first model.
This is a required entry.}

\item{seas_mdl2}{A \code{seas} object for a single series generated from the
\code{seasonal} package for the second model fit to the same series as
for \code{seas_mdl1}.
This is a required entry.}

\item{start_hist}{integer scalar; starting date for the history analysis.
Could be an array of length 2; will be converted to a scalar.
This is a required entry.}

\item{main_title}{Character string; main title of plot.
Default is  \code{'Differences in the Sum of Squared Forecast Errors'}.}

\item{name_mdl1}{Character string; Description of first model for use in the subtitle.
Default is \code{'Model 1'}.}

\item{name_mdl2}{Character string; Description of second model for use in the subtitle.
Default is \code{'Model 2'}.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{do_frame}{Logical scalar; indicates the plot will have a frame.
Default is the plot frame will be produced.}

\item{this_col}{Character array of length 2; color used for each forecast lag.
Default is colors from the \code{rcartocolor} palette \code{"Bold"}
\insertCite{rcartocolor}{blstinyplot}; this palette is extracted using the
\code{paletteer} package \insertCite{paletteer}{blstinyplot}.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_main_cex}{Numeric scalar; scaling for main plot title.
Default is the value of \code{this_plot_cex + 0.1}.}

\item{this_sub_cex}{Numeric scalar; scaling for plot subtitle.
Default is the value of \code{this_plot_cex}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(5.1, 2.1, 5.1, 0.5)}.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2.25}.}

\item{sub_title_line}{Integer scalar; position of main title of plot.
Default is \code{1}.}

\item{display_color_strip}{Logical scalar; indicates if the display color
will be stripped of trailing numbers.
Default is \code{TRUE}.}

\item{x_axis_label}{Character string; x-axis label for plot, if specified.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generate forecast history plot. Can be more than one series. If series not specified, print out error message and return NULL.
}
\description{
Generate forecast history plot, which compares the sum of squared forecast errors
for two models.
}
\details{
Version 1.3, 2026-05-18
}
\examples{
air_seas_mdl <- 
    seasonal::seas(AirPassengers, x11='', 
        slidingspans = '', transform.function = 'log',
        arima.model = '(0 1 1)(0 1 1)', 
        regression.aictest = NULL, outlier = NULL,
        forecast.maxlead = 36, 
        check.print = c( 'pacf', 'pacfplot' ),
        history.fstep = c(1, 12), history.estimates = 'fcst', 
        history.save = 'fcsterrors')
air_seas_mdl2 <- 
    seasonal::seas(AirPassengers, x11='', 
        slidingspans = '', transform.function = 'log',
        arima.model = '(0 1 1)(0 1 1)', 
        regression.variables = c("td"),
        forecast.maxlead = 36, 
        check.print = c( 'pacf', 'pacfplot' ),
        history.fstep = c(1, 12), history.estimates = 'fcst', 
        history.save = 'fcsterrors')
plot_fcst_history(air_seas_mdl, air_seas_mdl2, 
     start_hist = 1957.0,
     main_title = 'Differences in the SS Fcst Error for Airline Passengers',
     name_mdl1 = 'Airline model', 
     name_mdl2 = 'Airline model + regressors')
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
