% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fts.R
\name{plot_fts}
\alias{plot_fts}
\title{Final t-statistics for the outlier identification procedure plot}
\usage{
plot_fts(
  seas_obj = NULL,
  fts = NULL,
  this_cex = 0.5,
  start_plot = NULL,
  main_title = "Outlier T-Values",
  add_identified_otl = FALSE,
  do_grid = FALSE,
  color_otl = NULL,
  this_palette = "Dark2",
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_main_cex = NULL,
  this_sub_cex = NULL,
  this_axis_cex = NULL,
  this_mar = c(4, 4, 4, 0.5),
  main_title_line = 2.25,
  sub_title_line = 1,
  x_axis_label = NULL,
  y_axis_label = NULL,
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  plot_abs = TRUE,
  this_reset = TRUE
)
}
\arguments{
\item{seas_obj}{A \code{seas} object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{fts}{time series matrix containing final outlier t-statistics for all types
of outlier specified by the user.
This is a required entry.}

\item{this_cex}{Numeric scalar; sets the \code{cex} plotting parameter.
Default sets \code{cex = 0.5}.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{main_title}{Character string; main title of plot.
Default is  \code{'Outlier T-Values'}.}

\item{add_identified_otl}{Logical scalar; indicates if outlier plots will
include identified outliers.
Default is not including identified outliers.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{color_otl}{Character array of length 3; color used for different outliers,
with the order being \code{'ao','ls','tc'}.
Default is NULL.}

\item{this_palette}{Character string; default \code{RColorBrewer} palette
\insertCite{RColorBrewer}{blsplotGG}.
Deault is \code{"Dark2"}.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_main_cex}{Numeric scalar; scaling for main plot title.
Default is the value of \code{this_plot_cex + 0.1}.}

\item{this_sub_cex}{Numeric scalar; scaling for plot subtitle.
Default is the value of \code{this_plot_cex}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(4, 4, 4, 0.5)}.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2.25}.}

\item{sub_title_line}{Integer scalar; position of main title of plot.
Default is \code{1}.}

\item{x_axis_label}{Character string; x-axis label for plot, if specified.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{plot_abs}{Logical scalar; if \code{TRUE}, plot the maximum absolute value
of the t-stats. If \code{FALSE},
maintain the sign of the original t-statistic in the plot.
Default is \code{TRUE}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generates a plot of the final t-statistics from the automatic outlier
identification procedure.
}
\description{
Generates a plot of the final t-statistics for the outlier identification procedure.
}
\details{
Version 2.4, 2026-05-27
}
\examples{
air_seas_outlier <- 
    seasonal::seas(AirPassengers, 
        arima.model = '(0 1 1)(0 1 1)', 
        outlier.types = 'all',  
        outlier.save = "fts")
air_fts_good <- seasonal::series(air_seas_outlier, "fts")
plot_fts(air_seas_outlier, air_fts_good, 
         main_title = 'Outlier T-Values for Airline Passengers')
plot_fts(air_seas_outlier, air_fts_good, 
         main_title = 'Outlier T-Values for Airline Passengers',
         add_identified_outliers = TRUE)
plot_fts(air_seas_outlier, air_fts_good, plot_abs = FALSE,
         main_title = 'Outlier T-Values for Airline Passengers',
         add_identified_outliers = TRUE)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
