% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio.r
\name{plot_ratio}
\alias{plot_ratio}
\title{Ratio plot}
\usage{
plot_ratio(
  ratio_series = NULL,
  ratio_range = range(ratio_series),
  main_title = "Ratio Plot",
  main_title_line = 2,
  main_title_cex = NULL,
  ratio_mean = 1,
  ratio_color = "black",
  ratio_mean_color = "black",
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  x_axis_label = "Time",
  y_axis_label = NULL,
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_axis_cex = NULL,
  this_mar = c(4, 4, 4, 0.5),
  this_reset = TRUE
)
}
\arguments{
\item{ratio_series}{Time series of ratios/factors for which you want to
generate a high definition plot.
This is a required entry.}

\item{ratio_range}{Range of values you wish the plot to be plotted over.
Default is range of the series.}

\item{main_title}{Title for the plot.
Default is character string \code{'Ratio Plot'}.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2}.}

\item{main_title_cex}{Numeric scalar; scaling for main title of plot.
Default is \code{this_plot_cex + 0.25}.}

\item{ratio_mean}{Assumed mean value for the ratio.
Default is \code{1.0}}

\item{ratio_color}{Color used for lines in ratio plot.
Default is \code{'black'}.}

\item{ratio_mean_color}{Color used for mean line in ratio plot.
Default is \code{'black'}.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{x_axis_label}{Character string; x-axis label for plot.
Default is \code{'Time'}.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(4,4,4,0.5)}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generates a high definition plot of rations centered on one,
by default.
}
\description{
Generates a high-definition plot around a reference line other than zero.
}
\details{
Version 1.5, 2026-04-02
}
\examples{
air_seas <-   
    seasonal::seas(AirPassengers,  
        transform.function= 'log', 
        arima.model = '(0 1 1)(0 1 1)',   
        seats.save = 's10')
air_sf <- seasonal::series(air_seas, 's10')
plot_ratio(air_sf,  
     main_title = 'SEATS seasonal for Airline Passenger',  
     ratio_color = 'darkblue')
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
