% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_resid.R
\name{plot_resid}
\alias{plot_resid}
\title{Residual plot}
\usage{
plot_resid(
  seas_obj = NULL,
  main_title = "ARIMA Residuals",
  main_title_line = 2.75,
  main_title_cex = 1,
  series_name = NULL,
  x_axis_label = NULL,
  y_axis_label = NULL,
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  do_grid = TRUE,
  draw_recess = FALSE,
  recess_start = NULL,
  recess_col = NULL,
  recess_sub = TRUE,
  this_trans = TRUE,
  use_ratio = FALSE,
  this_col = "green",
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_axis_cex = NULL,
  this_sub_cex = 0.5,
  this_mar = c(4, 4, 4, 0.5),
  sub_title_one_line = 1.5,
  sub_title_two_line = 0.5,
  this_reset = TRUE
)
}
\arguments{
\item{seas_obj}{A \code{seas} object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{main_title}{Character string; main title of plot.
Default is  \code{'ARIMA Residuals'}.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2.75}.}

\item{main_title_cex}{Numeric scalar; scaling for main title of plot.
Default is \code{1.25}.}

\item{series_name}{Character scalar; name of the time series used in
\code{seas_obj}. Used only if specified.}

\item{x_axis_label}{Character string; x-axis label for plot, if specified.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is grid lines plotted.}

\item{draw_recess}{Logical scalar; indicates if certain plots will
have shaded areas for NBER recession dates \insertCite{nber}{blstinyplot}.
Default is recession shading not plotted.}

\item{recess_start}{numeric matrix; Rows of dates for additional recession
starting and ending dates.
Default is not to add recession dates.}

\item{recess_col}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession is
produced for this plot.
Default is x-axis label is produced}

\item{this_trans}{Logical scalar; indicates if the adjustment was done
with a log transform. Default is TRUE.}

\item{use_ratio}{Logical scalar; indicates if plots of seasonal factors,
irregular, and residuals are done as ratio plots.
Default has these plots as time series line plots.}

\item{this_col}{Character string; color used for residuals.
Default is \code{'green'}.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_sub_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{0.5}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(4,4,4,0,5)}.}

\item{sub_title_one_line}{Integer scalar; position of furst subtitle of plot.
Default is \code{1.5}.}

\item{sub_title_two_line}{Integer scalar; position of main title of plot.
Default is \code{0.5}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generates a plot of the regARIMA residuals with diagnostic information
in the sub-headers.
}
\description{
Generates a plot of the regARIMA residuals with diagnostic information
}
\details{
Version 1.4, 2026-05-21
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, arima.model = '(0 1 1)(0 1 1)')
plot_resid(air_seas, 
           main_title = 'ARIMA Residuals for Airline Passengers', 
           use_ratio = TRUE, this_col='darkblue')
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
