% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_series.r
\name{plot_series}
\alias{plot_series}
\title{Plot individual series.}
\usage{
plot_series(
  this_series = NULL,
  main_title = NULL,
  main_title_line = 2.75,
  main_title_cex = 1.25,
  x_axis_label = NULL,
  y_axis_label = NULL,
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  y_limit = NULL,
  start_plot = NULL,
  do_grid = FALSE,
  draw_recess = FALSE,
  recess_start = NULL,
  recess_col = "lightgrey",
  recess_sub = TRUE,
  this_trans = TRUE,
  use_ratio = FALSE,
  this_col = "grey",
  this_plot_type = "l",
  this_line_type = 1,
  this_line_width = 1,
  this_point_type = NULL,
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_axis_cex = NULL,
  this_mar = c(4, 4, 4, 0.5),
  this_reset = TRUE
)
}
\arguments{
\item{this_series}{Numeric vector; time series object to be plotted.
This is a required entry.}

\item{main_title}{Character string; main title of plot.
Default is no title.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2.75}.}

\item{main_title_cex}{Numeric scalar; scaling for main title of plot.
Default is \code{1.25}.}

\item{x_axis_label}{Character string; x-axis label for plot, if specified.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_limit}{Numeric vector of length 2; Range of values you wish
the plot to be plotted over.
Default is range of \code{this_series}.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{do_grid}{Logical scalar; indicates if certain plots will have
grid lines. Default is no grid lines.}

\item{draw_recess}{Logical scalar; indicates if certain plots will
have shaded areas for NBER recession dates \insertCite{nber}{blstinyplot}.
Default is no recession shading.}

\item{recess_start}{numeric matrix; Rows of dates for additional
recession starting and ending dates.
Default is not to add recession dates.}

\item{recess_col}{Character string; color used for shading of
recession region. Default is \code{"lightgrey"}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for
recession is produced for this plot.
Default is x-axis label is produced}

\item{this_trans}{Logical scalar; indicates if the adjustment was done
with a log transform. Default is TRUE.}

\item{use_ratio}{Logical scalar; indicates if plots of seasonal factors,
irregular, and residuals are done as ratio plots.
Default has these plots as time series line plots.}

\item{this_col}{Character string; color used for series.
Default is \code{'grey'}.}

\item{this_plot_type}{Character string; type of plot produced.
Typical values are \code{"l"} (lines only), \code{"p"} (points only),
\code{"b"} (both points and lines), \code{"o"} (overplotted points
and lines), \code{"n"} (no plotting).
Default is "l".}

\item{this_line_type}{Integer scalar; line type used for series.
Default is \code{1}.}

\item{this_line_width}{Integer scalar; line width used for series.
Default is \code{1}.}

\item{this_point_type}{Integer scalar; point type used for series.
Default is no points plotted.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(4,4,4,0,5)}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generate plot of user-specified series. Can be first in a series
of plots, with other lines or points added after calling this routine. If series not specified, print out error message and return NULL.
}
\description{
Generate plot of user-specified series.
}
\details{
Version 2.0, 2026-04-08
}
\examples{
air_seas <- 
    seasonal::seas(AirPassengers, 
        arima.model = "(0 1 1)(0 1 1)",
        outlier.types = "all", 
        x11 = "",
        forecast.maxlead = 36)
air_sa <- seasonal::final(air_seas)
plot_series(air_sa, 
     y_axis_label = "AirPass", 
     do_grid = TRUE,
     draw_recess = TRUE, 
     this_col = "black",
     start_plot = c(1958,1), 
     this_point_type = 1,
     main_title = "X-11 Seasonal Adjustment for Airline Passengers",
     main_title_line = 1.5,
     main_title_cex = 1.0, 
     this_reset = TRUE)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
