% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sf.R
\name{plot_sf}
\alias{plot_sf}
\title{Seasonal factor (and the SI-ratios) plot grouped by month/quarter}
\usage{
plot_sf(
  seas_obj = NULL,
  this_table = NULL,
  add_si = FALSE,
  main_title = "Seasonal Sub-Plots",
  sub_title = NULL,
  y_label = NULL,
  y_limit = NULL,
  x_label = " ",
  this_col = NULL,
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_main_cex = NULL,
  this_axis_cex = NULL,
  this_mar = c(2.1, 1.5, 3.1, 0.5),
  num_facet_rows = NULL,
  do_grid = FALSE,
  this_facet_bg = "grey90",
  this_facet_cex = 0.6,
  this_tinytheme = "minimal",
  this_reset = TRUE
)
}
\arguments{
\item{seas_obj}{A seas object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{this_table}{Character string; X-13ARIMA-SEATS table name or abbreviation
(such as \code{d10} or \code{s10}). Valid table names for this argument
can be found in the documentation for the \code{series}
function of the \code{seasonal} package \insertCite{Sax2018}{blstinyplot}.
If not a valid table name, the function will print an error message and
return a \code{NULL}. This entry is required.}

\item{add_si}{Logical scalar; indicates if seasonal factor plots will include
the SI ratios and unmodified SI ratios for X-11 seasonal adjustments.
Default is not including SI ratios (\code{add_si = FALSE}).}

\item{main_title}{Character string; main title of plot.
Default is \code{'Seasonal Sub-Plots'}.}

\item{sub_title}{Character string; a subtitle for the plot.
Default is \code{NULL}, which generates a subtitle
which denotes the series that are plotted.}

\item{y_label}{Character string; y-axis label for plot, if specified.}

\item{y_limit}{Numeric vector of length 2; Range of values to be plotted over.
Default is range of the seasonal factors.}

\item{x_label}{Character string; label for x-axis of plot.
Default is a blank x-axis.}

\item{this_col}{Character array of length 2 or 4; color used for
seasonal factors, SI-ratios, Modified SI-ratios, and seasonal means.
Default is generated from the \code{RColorBrewer} palette \code{"Dark2"},
with \code{"darkgrey"} added for the seasonal mean line.
For a listing of R color codes,
see \insertCite{rchartsColors;textual}{blsplotGG}.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_main_cex}{Numeric scalar; scaling for main plot title.
Default is the value of \code{this_plot_cex + 0.1}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(2.1, 1.5, 3.1, 0.5)}.}

\item{num_facet_rows}{Integer scalar; number of rows in the facet plot.
Default is generated from the series frequency (\code{2} for
monthly series, 1 otherwise).}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{this_facet_bg}{Character scalar; Color used for the facet labels.
Default is \code{"grey90"}.}

\item{this_facet_cex}{Integer scalar; size used for the facet labels.
Default is \code{0.6}.}

\item{this_tinytheme}{Character scalar; setting for the \code{tinythere} function
from the \code{tinyplot} package \insertCite{tinyplot}{blstinyplot}.
Default is \code{"minimal"}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{TRUE}.}
}
\value{
Generates a special plot of the seasonal factors (and the SI-ratios)
grouped by month/quarter.
}
\description{
Generates a special plot of the seasonal factors (and the SI-ratios) grouped by month/quarter
}
\details{
Version 1.3.1, 2026-05-14
}
\examples{
air_seas <- seasonal::seas(AirPassengers, arima.model = '(0 1 1)(0 1 1)', 
                           x11='', x11.save = c("d8", "d9"))
plot_sf(air_seas, add_si = TRUE, 
        main_title = 'Air Passengers Seasonal Sub-Plots',
        sub_title = "SF = Cyan, SI = Orange, ModSI = Purple, Mean = Gray",
        do_grid = TRUE)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
