% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_table.r
\name{plot_table}
\alias{plot_table}
\title{Plot table from X-13ARIMA-SEATS seasonal adjustment run.}
\usage{
plot_table(
  seas_obj = NULL,
  this_table = NULL,
  main_title = NULL,
  main_title_line = 2,
  main_title_cex = 1.25,
  x_axis_label = NULL,
  y_axis_label = NULL,
  x_axis_main_ticks = NULL,
  y_axis_main_ticks = NULL,
  y_limit = NULL,
  start_plot = NULL,
  do_grid = FALSE,
  draw_recess = FALSE,
  recess_start = NULL,
  recess_col = NULL,
  recess_sub = TRUE,
  add_otl = FALSE,
  use_ratio = FALSE,
  add_sub_title = FALSE,
  sub_title_line = 1,
  sub_title_cex = NULL,
  this_line_type = NULL,
  this_line_width = 1,
  this_point_type = NULL,
  this_col = "grey",
  otl_col = NULL,
  this_plot_cex = 0.8,
  this_lab_cex = NULL,
  this_axis_cex = NULL,
  this_mar = c(4, 4, 4, 0.5),
  this_reset = FALSE
)
}
\arguments{
\item{seas_obj}{A \code{seas} object for a single series generated from the
\code{seasonal} package \insertCite{seasonal2018}{blstinyplot}.
This is a required entry.}

\item{this_table}{Character string; X-13ARIMA-SEATS table name or abbreviation.
If not a valid table name, the function will print an error message and return a \code{NULL}.
This entry is required.}

\item{main_title}{Character string; main title of plot.
Default is \code{'Cumulative periodogram'}.}

\item{main_title_line}{Integer scalar; position of main title of plot.
Default is \code{2.75}.}

\item{main_title_cex}{Numeric scalar; scaling for main title of plot.
Default is \code{this_plot_cex + 0.1}.}

\item{x_axis_label}{Character string; x-axis label for plot, if specified.}

\item{y_axis_label}{Character string; y-axis label for plot, if specified.}

\item{x_axis_main_ticks}{Numeric vector; X-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_axis_main_ticks}{Numeric vector; Y-axis tick marks.
Default is determined by the \code{axis} function.}

\item{y_limit}{Numeric vector of length 2; Range of values you wish
the plot to be plotted over.
Default is range of the series specified.}

\item{start_plot}{Integer vector of length 2; Starting date for plot.
Default is starting date for the time series.}

\item{do_grid}{Logical scalar; indicates if certain plots will have grid lines.
Default is no grid lines.}

\item{draw_recess}{Logical scalar; indicates if certain plots will have
shaded areas for NBER recession dates \insertCite{nber}{blstinyplot}.
Default is no recession shading.}

\item{recess_start}{numeric matrix; Rows of dates for additional recession
starting and ending dates.
Default is not to add recession dates.}

\item{recess_col}{Character string; color used for shading of recession region.
Default is \code{'lightgrey'}.}

\item{recess_sub}{Logical scalar; indicates if x-axis label for recession
is produced for this plot.
Default is x-axis label is produced}

\item{add_otl}{Logical scalar; indicates if lines for identified outliers
are included in series plots.
Default is not including lines for identified outliers.}

\item{use_ratio}{Logical scalar; indicates if plots of seasonal factors,
irregular, and residuals are done as ratio plots.
Default has these plots as time series line plots.}

\item{add_sub_title}{Logical scalar; indicates if plots will include
subtitles denoting what series are plotted.
Default is not including subheaders.}

\item{sub_title_line}{Integer scalar; position of main title of plot.
Default is \code{1}.}

\item{sub_title_cex}{Numeric scalar; scaling for main title of plot.
Default is \code{this_plot_cex - 0.1}.}

\item{this_line_type}{Integer scalar; indicates line type of plot.
Default is \code{1}.}

\item{this_line_width}{Integer scalar; line width used for series.
Default is \code{1}.}

\item{this_point_type}{Integer scalar; point type used for series.
Default is no points plotted.}

\item{this_col}{Character scalar; color used for plot.
Default is \code{'grey'}.}

\item{otl_col}{Character array of length 6; color used for different
outliers, with the order being \code{'ao'}, \code{'ls'},
\code{'tc'}, \code{'so'}, \code{'rp'}, \code{'tls'}.
Default is colors from the \code{RColorBrewer} palette \code{"Paired"}
\insertCite{RColorBrewer}{blstinyplot}.}

\item{this_plot_cex}{Numeric scalar; scaling for the plot itself.
Default is \code{0.8}.}

\item{this_lab_cex}{Numeric scalar; scaling for plot labels.
Default is the value of \code{this_plot_cex}.}

\item{this_axis_cex}{Numeric scalar; scaling for plot axis.
Default is the value of \code{this_plot_cex}.}

\item{this_mar}{Numeric vector; set margins for the plot.
Default is \code{c(4,4,4,0,5)}.}

\item{this_reset}{Logical scalar; if TRUE, the values of \code{par} are reset.
Default is \code{FALSE}.}
}
\value{
Generate plot of user-specified series. If series not specified,
print out error message and return NULL.
}
\description{
Generate plot of user-specified table from the output of a \code{seas} object
generated by the \code{seasonal} package.
}
\details{
Version 1.5, 2026-04-08
}
\examples{
air_seas <- 
   seasonal::seas(AirPassengers, 
       arima.model = "(0 1 1)(0 1 1)",
       x11="", x11.save = c("e18"))
blstinyplot::plot_table(air_seas, "d11", 
         y_axis_label = "AirPassengers", 
         main_title = "X-11 Seasonal Adjustment of Airline Passengers",
         main_title_cex = 1.0, 
         add_sub_title = TRUE, 
         do_grid = TRUE, 
         draw_recess = TRUE, 
         use_ratio = FALSE,
         add_otl = TRUE,
         this_col = "black")
blstinyplot::plot_table(air_seas, "e18", 
         y_axis_label = "AirPassengers", 
         main_title = "Adjustment Factors of Airline Passengers",
         main_title_cex = 1.0, 
         add_sub_title = TRUE, 
         do_grid = TRUE, 
         draw_recess = TRUE, 
         use_ratio = TRUE,
         add_otl = TRUE,
         this_col = "darkblue")
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
