% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_par.r
\name{reset_par}
\alias{reset_par}
\title{Reset \code{par()}}
\usage{
reset_par(old_par = NULL)
}
\arguments{
\item{old_par}{List object; old graphical parameters saved from \code{par}.
This is a required entry.}
}
\value{
returns reset graphics parameters
}
\description{
Reset graphics parameters for plots; taken from \insertCite{resetPar;textual}{blstinyplot}.
}
\details{
Version 1.1, 2026-04-07
}
\examples{
old.par <- par(no.readonly = TRUE)
par(mar=c(5.1, 3.1, 4.1, 1.1), mfrow=c(2,2))
xt_names <- names(xt_data_list)
for (i in 1:4) {
     plot(xt_data_list[[i]], main = xt_names[i], type="l")
}
reset_par(old.par)
}
\references{
\insertAllCited{}
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
