% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_day_and_month.r
\name{set_day_and_month}
\alias{set_day_and_month}
\title{generate date object from days and year}
\usage{
set_day_and_month(day_number, target_year)
}
\arguments{
\item{day_number}{Integer scalar, number of days from the start of the year.
This is a required entry.}

\item{target_year}{Integer scalar, year of the date object.
This is a required entry.}
}
\value{
Date object determined by the input arguments.
}
\description{
Generate a date object from the number of days from the beginning of the year and the target year.
}
\details{
Version 1.1, 1/7/2026
}
\examples{
this_start <- set_day_and_month(1, 2001)
this_end   <- set_day_and_month(350, 2025)
}
\author{
Brian C. Monsell, \email{monsell.brian@bls.gov} or \email{bcmonsell@gmail.com}
}
